+++ atm_dyn_iconam/tests/test_mo_solve_nonhydro_stencil_20.py
@@ -22,6 +22,7 @@ from icon4py.atm_dyn_iconam.mo_solve_nonhydro_stencil_20_itir import (
     mo_solve_nonhydro_stencil_20_itir,
 )
 
+from gt4py.next.ffront.fbuiltins import int32
 from icon4py.common.dimension import CellDim, E2CDim, ECDim, EdgeDim, KDim
 from icon4py.testutils.simple_mesh import SimpleMesh
 from icon4py.testutils.utils import flatten_first_two_dims, random_field, zero_field
@@ -76,7 +77,7 @@ def test_mo_solve_nonhydro_stencil_20():
     inv_dual_edge_length = random_field(mesh, EdgeDim)
     z_exner_ex_pr = random_field(mesh, CellDim, KDim)
     zdiff_gradp = random_field(mesh, EdgeDim, E2CDim, KDim)
-    ikoffset = zero_field(mesh, EdgeDim, E2CDim, KDim, dtype=int)
+    ikoffset = zero_field(mesh, EdgeDim, E2CDim, KDim, dtype=int32)
 
     rng = np.random.default_rng()
     for k in range(mesh.k_level):
